integrate java generator
