+++ ci-script/src/builders/java.rs
@@ -54,7 +54,7 @@ impl BindingBuilder for JavaBindingBuilder {
     }
 
     fn generate(&mut self, _is_packaging: bool, _generate_doxygen: bool) {
-        let config = java_oo_bindgen::JavaBindgenConfig {
+        let config = oo_bindgen::backend::java::JavaBindgenConfig {
             java_output_dir: self.java_build_dir(),
             ffi_name: self.settings.ffi_name,
             ffi_path: self.settings.ffi_path.to_owned(),
@@ -71,7 +71,7 @@ impl BindingBuilder for JavaBindingBuilder {
         logged::create_dir_all(&build_dir).unwrap();
 
         // Generate the Java code
-        java_oo_bindgen::generate_java_bindings(&self.settings.library, &config).unwrap();
+        oo_bindgen::backend::java::generate_java_bindings(&self.settings.library, &config).unwrap();
     }
 
     fn build(&mut self) {
