--- generators/java-oo-bindgen/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "java-oo-bindgen"
-version = "0.5.1"
-authors = ["Step Function I/O LLC <info@stepfunc.io>"]
-edition = "2021"
-
-[dependencies]
-oo-bindgen = { path = "../../oo-bindgen" }
-dunce = "1.0"
-pathdiff = "0.2"
-tracing = "^0.1"
