+++ oo-bindgen/src/backend/c/mod.rs
@@ -81,7 +81,7 @@ pub fn generate_c_package(lib: &Library, config: &CBindgenConfig) -> FormattingR
 
 fn generate_doxygen(lib: &Library, config: &CBindgenConfig) -> FormattingResult<()> {
     // Copy doxygen awesome in target directory
-    let doxygen_awesome = include_str!("../../../../generators/doxygen-awesome.css");
+    let doxygen_awesome = include_str!("../../../static/doxygen-awesome.css");
     fs::write(
         config.output_dir.join("doxygen-awesome.css"),
         doxygen_awesome,
