+++ oo-bindgen/src/backend/java/rust/conversion/convertible_to_jni.rs
@@ -1,4 +1,4 @@
-use oo_bindgen::model::*;
+use crate::model::*;
 
 pub(crate) trait MaybeConvertibleToJni {
     /// Possible conversion from JNI argument type to Rust type
