+++ tests/foo-ffi-java/Cargo.toml
@@ -13,4 +13,4 @@ foo-ffi = { path = "../foo-ffi" }
 
 [build-dependencies]
 foo-schema = { path = "../foo-schema" }
-java-oo-bindgen = { path = "../../generators/java-oo-bindgen" }
+oo-bindgen = { path = "../../oo-bindgen" }
