+++ tests/foo-ffi-java/build.rs
@@ -1,5 +1,3 @@
-use java_oo_bindgen::generate_java_ffi;
-use java_oo_bindgen::JniBindgenConfig;
 use std::path::{Path, PathBuf};
 
 fn main() {
@@ -9,7 +7,7 @@ fn main() {
     // normally you'd never want to write files here, but this crate isn't used as a dependency
     let out_path: PathBuf = Path::new(&std::env::var_os("OUT_DIR").unwrap()).join("jni.rs");
 
-    let config = JniBindgenConfig {
+    let config = oo_bindgen::backend::java::JniBindgenConfig {
         group_id: "io.stepfunc".to_string(),
         ffi_name: "foo_ffi".to_string(),
     };
@@ -20,7 +18,7 @@ fn main() {
             std::process::exit(-1);
         }
         Ok(lib) => {
-            generate_java_ffi(&out_path, &lib, &config).unwrap();
+            oo_bindgen::backend::java::generate_java_ffi(&out_path, &lib, &config).unwrap();
         }
     }
 }
